refactor: initial refactor to support milli migration
