+++ packages/mimir/lib/src/bridge_generated.dart
@@ -19,13 +19,25 @@ abstract class EmbeddedMilli {
 
   FlutterRustBridgeTaskConstMeta get kEnforceBindingConstMeta;
 
+  /// Ensures an instance of milli (represented by just a directory) is initialized
+  Future<void> ensureInstanceInitialized(
+      {required String instaceDir, dynamic hint});
+
+  FlutterRustBridgeTaskConstMeta get kEnsureInstanceInitializedConstMeta;
+
+  /// Ensures a milli index is initialized
+  Future<void> ensureIndexInitialized(
+      {required String instanceDir, required String indexName, dynamic hint});
+
+  FlutterRustBridgeTaskConstMeta get kEnsureIndexInitializedConstMeta;
+
   /// Adds the given list of documents to the specified milli index
   ///
   /// Replaces documents that already exist in the index based on document ids.
   Future<void> addDocuments(
       {required String instanceDir,
       required String indexName,
-      required List<String> jsonDocuments,
+      required List<String> documents,
       dynamic hint});
 
   FlutterRustBridgeTaskConstMeta get kAddDocumentsConstMeta;
@@ -49,7 +61,7 @@ abstract class EmbeddedMilli {
   Future<void> setDocuments(
       {required String instanceDir,
       required String indexName,
-      required List<String> jsonDocuments,
+      required List<String> documents,
       dynamic hint});
 
   FlutterRustBridgeTaskConstMeta get kSetDocumentsConstMeta;
@@ -256,20 +268,59 @@ class EmbeddedMilliImpl implements EmbeddedMilli {
         argNames: [],
       );
 
+  Future<void> ensureInstanceInitialized(
+      {required String instaceDir, dynamic hint}) {
+    var arg0 = _platform.api2wire_String(instaceDir);
+    return _platform.executeNormal(FlutterRustBridgeTask(
+      callFfi: (port_) =>
+          _platform.inner.wire_ensure_instance_initialized(port_, arg0),
+      parseSuccessData: _wire2api_unit,
+      constMeta: kEnsureInstanceInitializedConstMeta,
+      argValues: [instaceDir],
+      hint: hint,
+    ));
+  }
+
+  FlutterRustBridgeTaskConstMeta get kEnsureInstanceInitializedConstMeta =>
+      const FlutterRustBridgeTaskConstMeta(
+        debugName: "ensure_instance_initialized",
+        argNames: ["instaceDir"],
+      );
+
+  Future<void> ensureIndexInitialized(
+      {required String instanceDir, required String indexName, dynamic hint}) {
+    var arg0 = _platform.api2wire_String(instanceDir);
+    var arg1 = _platform.api2wire_String(indexName);
+    return _platform.executeNormal(FlutterRustBridgeTask(
+      callFfi: (port_) =>
+          _platform.inner.wire_ensure_index_initialized(port_, arg0, arg1),
+      parseSuccessData: _wire2api_unit,
+      constMeta: kEnsureIndexInitializedConstMeta,
+      argValues: [instanceDir, indexName],
+      hint: hint,
+    ));
+  }
+
+  FlutterRustBridgeTaskConstMeta get kEnsureIndexInitializedConstMeta =>
+      const FlutterRustBridgeTaskConstMeta(
+        debugName: "ensure_index_initialized",
+        argNames: ["instanceDir", "indexName"],
+      );
+
   Future<void> addDocuments(
       {required String instanceDir,
       required String indexName,
-      required List<String> jsonDocuments,
+      required List<String> documents,
       dynamic hint}) {
     var arg0 = _platform.api2wire_String(instanceDir);
     var arg1 = _platform.api2wire_String(indexName);
-    var arg2 = _platform.api2wire_StringList(jsonDocuments);
+    var arg2 = _platform.api2wire_StringList(documents);
     return _platform.executeNormal(FlutterRustBridgeTask(
       callFfi: (port_) =>
           _platform.inner.wire_add_documents(port_, arg0, arg1, arg2),
       parseSuccessData: _wire2api_unit,
       constMeta: kAddDocumentsConstMeta,
-      argValues: [instanceDir, indexName, jsonDocuments],
+      argValues: [instanceDir, indexName, documents],
       hint: hint,
     ));
   }
@@ -277,7 +328,7 @@ class EmbeddedMilliImpl implements EmbeddedMilli {
   FlutterRustBridgeTaskConstMeta get kAddDocumentsConstMeta =>
       const FlutterRustBridgeTaskConstMeta(
         debugName: "add_documents",
-        argNames: ["instanceDir", "indexName", "jsonDocuments"],
+        argNames: ["instanceDir", "indexName", "documents"],
       );
 
   Future<void> deleteDocuments(
@@ -327,17 +378,17 @@ class EmbeddedMilliImpl implements EmbeddedMilli {
   Future<void> setDocuments(
       {required String instanceDir,
       required String indexName,
-      required List<String> jsonDocuments,
+      required List<String> documents,
       dynamic hint}) {
     var arg0 = _platform.api2wire_String(instanceDir);
     var arg1 = _platform.api2wire_String(indexName);
-    var arg2 = _platform.api2wire_StringList(jsonDocuments);
+    var arg2 = _platform.api2wire_StringList(documents);
     return _platform.executeNormal(FlutterRustBridgeTask(
       callFfi: (port_) =>
           _platform.inner.wire_set_documents(port_, arg0, arg1, arg2),
       parseSuccessData: _wire2api_unit,
       constMeta: kSetDocumentsConstMeta,
-      argValues: [instanceDir, indexName, jsonDocuments],
+      argValues: [instanceDir, indexName, documents],
       hint: hint,
     ));
   }
@@ -345,7 +396,7 @@ class EmbeddedMilliImpl implements EmbeddedMilli {
   FlutterRustBridgeTaskConstMeta get kSetDocumentsConstMeta =>
       const FlutterRustBridgeTaskConstMeta(
         debugName: "set_documents",
-        argNames: ["instanceDir", "indexName", "jsonDocuments"],
+        argNames: ["instanceDir", "indexName", "documents"],
       );
 
   Future<String?> getDocument(
@@ -905,17 +956,56 @@ class EmbeddedMilliWire implements FlutterRustBridgeWireBase {
   late final _wire_enforce_binding =
       _wire_enforce_bindingPtr.asFunction<void Function(int)>();
 
+  void wire_ensure_instance_initialized(
+    int port_,
+    ffi.Pointer<wire_uint_8_list> instace_dir,
+  ) {
+    return _wire_ensure_instance_initialized(
+      port_,
+      instace_dir,
+    );
+  }
+
+  late final _wire_ensure_instance_initializedPtr = _lookup<
+          ffi.NativeFunction<
+              ffi.Void Function(ffi.Int64, ffi.Pointer<wire_uint_8_list>)>>(
+      'wire_ensure_instance_initialized');
+  late final _wire_ensure_instance_initialized =
+      _wire_ensure_instance_initializedPtr
+          .asFunction<void Function(int, ffi.Pointer<wire_uint_8_list>)>();
+
+  void wire_ensure_index_initialized(
+    int port_,
+    ffi.Pointer<wire_uint_8_list> instance_dir,
+    ffi.Pointer<wire_uint_8_list> index_name,
+  ) {
+    return _wire_ensure_index_initialized(
+      port_,
+      instance_dir,
+      index_name,
+    );
+  }
+
+  late final _wire_ensure_index_initializedPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Void Function(ffi.Int64, ffi.Pointer<wire_uint_8_list>,
+              ffi.Pointer<wire_uint_8_list>)>>('wire_ensure_index_initialized');
+  late final _wire_ensure_index_initialized =
+      _wire_ensure_index_initializedPtr.asFunction<
+          void Function(int, ffi.Pointer<wire_uint_8_list>,
+              ffi.Pointer<wire_uint_8_list>)>();
+
   void wire_add_documents(
     int port_,
     ffi.Pointer<wire_uint_8_list> instance_dir,
     ffi.Pointer<wire_uint_8_list> index_name,
-    ffi.Pointer<wire_StringList> json_documents,
+    ffi.Pointer<wire_StringList> documents,
   ) {
     return _wire_add_documents(
       port_,
       instance_dir,
       index_name,
-      json_documents,
+      documents,
     );
   }
 
@@ -980,13 +1070,13 @@ class EmbeddedMilliWire implements FlutterRustBridgeWireBase {
     int port_,
     ffi.Pointer<wire_uint_8_list> instance_dir,
     ffi.Pointer<wire_uint_8_list> index_name,
-    ffi.Pointer<wire_StringList> json_documents,
+    ffi.Pointer<wire_StringList> documents,
   ) {
     return _wire_set_documents(
       port_,
       instance_dir,
       index_name,
-      json_documents,
+      documents,
     );
   }
