+++ packages/mimir/native/src/lib.rs
@@ -1,2 +1,3 @@
 mod api;
 mod bridge_generated; /* AUTO INJECTED BY flutter_rust_bridge. This line may not be accurate, and you can change it according to your needs. */
+mod embedded_milli;
